cast as int

+++ src/Model.py
@@ -18,7 +18,7 @@ class Model:
         self.__d_time: float = 1e-3
         """[s] fixed simulation time step"""
 
-        self.__num_time_steps: int = floor(config.get_end_time() / self.__d_time)
+        self.__num_time_steps: int = int(floor(config.get_end_time() / self.__d_time))
         """[] number of timesteps in the simulation"""
 
         # number of steps between snapshots
